+++ packages/isar_core_ffi/src/filter.rs
@@ -6,7 +6,7 @@ use isar_core::core::{
 use itertools::Itertools;
 
 #[no_mangle]
-pub unsafe extern "C" fn isar_filter_is_null(property_index: u32) -> *const Filter {
+pub unsafe extern "C" fn isar_filter_is_null(property_index: u16) -> *const Filter {
     let filter = Filter::Condition(FilterCondition::new(
         property_index,
         ConditionType::IsNull,
@@ -17,7 +17,7 @@ pub unsafe extern "C" fn isar_filter_is_null(property_index: u32) -> *const Filt
 }
 
 #[no_mangle]
-pub unsafe extern "C" fn isar_filter_list_is_empty(property_index: u32) -> *const Filter {
+pub unsafe extern "C" fn isar_filter_list_is_empty(property_index: u16) -> *const Filter {
     let filter = Filter::Condition(FilterCondition::new(
         property_index,
         ConditionType::ListIsEmpty,
@@ -29,7 +29,7 @@ pub unsafe extern "C" fn isar_filter_list_is_empty(property_index: u32) -> *cons
 
 #[no_mangle]
 pub unsafe extern "C" fn isar_filter_equal_to(
-    property_index: u32,
+    property_index: u16,
     value: *mut IsarValue,
     case_sensitive: bool,
 ) -> *const Filter {
@@ -49,7 +49,7 @@ pub unsafe extern "C" fn isar_filter_equal_to(
 
 #[no_mangle]
 pub unsafe extern "C" fn isar_filter_greater_than(
-    property_index: u32,
+    property_index: u16,
     value: *mut IsarValue,
     include: bool,
     case_sensitive: bool,
@@ -76,7 +76,7 @@ pub unsafe extern "C" fn isar_filter_greater_than(
 
 #[no_mangle]
 pub unsafe extern "C" fn isar_filter_less_than(
-    property_index: u32,
+    property_index: u16,
     value: *mut IsarValue,
     include: bool,
     case_sensitive: bool,
@@ -103,7 +103,7 @@ pub unsafe extern "C" fn isar_filter_less_than(
 
 #[no_mangle]
 pub unsafe extern "C" fn isar_filter_between(
-    property_index: u32,
+    property_index: u16,
     lower: *mut IsarValue,
     include_lower: bool,
     upper: *mut IsarValue,
@@ -143,7 +143,7 @@ pub unsafe extern "C" fn isar_filter_between(
 
 #[no_mangle]
 pub unsafe extern "C" fn isar_filter_string_starts_with(
-    property_index: u32,
+    property_index: u16,
     value: *mut IsarValue,
     case_sensitive: bool,
 ) -> *const Filter {
@@ -159,7 +159,7 @@ pub unsafe extern "C" fn isar_filter_string_starts_with(
 
 #[no_mangle]
 pub unsafe extern "C" fn isar_filter_string_ends_with(
-    property_index: u32,
+    property_index: u16,
     value: *mut IsarValue,
     case_sensitive: bool,
 ) -> *const Filter {
@@ -175,7 +175,7 @@ pub unsafe extern "C" fn isar_filter_string_ends_with(
 
 #[no_mangle]
 pub unsafe extern "C" fn isar_filter_string_contains(
-    property_index: u32,
+    property_index: u16,
     value: *mut IsarValue,
     case_sensitive: bool,
 ) -> *const Filter {
@@ -191,7 +191,7 @@ pub unsafe extern "C" fn isar_filter_string_contains(
 
 #[no_mangle]
 pub unsafe extern "C" fn isar_filter_string_matches(
-    property_index: u32,
+    property_index: u16,
     value: *mut IsarValue,
     case_sensitive: bool,
 ) -> *const Filter {
